Add PHP MySQL extensions
The PHP8-compatible versions of WPBrowser & Codeception, that we use to
run integration/acceptance tests on plugins like
dxw-Comment-Notifications and dxw-Digest, have changed the way they
communicate with databases. They now require these additional extensions
to be able to do so.

Adding them to our local image means we will be able to upgrade those
tests to work on PHP8.